Moved dto package out of the web package
