+++ src/main/java/com/sda/eventapp/web/mapper/EventMapper.java
@@ -1,7 +1,7 @@
 package com.sda.eventapp.web.mapper;
 
+import com.sda.eventapp.dto.EventWithBasicData;
 import com.sda.eventapp.model.Event;
-import com.sda.eventapp.web.dto.EventWithBasicData;
 import com.sda.eventapp.web.mvc.form.CreateEventForm;
 
 import java.util.List;
